tickets #397, #416, #311, #379
#397 - Changed bubble to a pretty green
#416 - Changed "delete" to "remove"
#311 - fixed block quote (adding spacing left/right). tested ul and ol's (added more spacing).
tested all other formatting options (bold, italic, underline, strikethrough, syntax highlighting, subscript, superscript) - no changes.
#379 - this ticket required quite a bit of finagling. I did some UI and would appreciate comments back on the solution. I ended up moving all the buttons from the sidebar into the title bar, as it doesn't make sense (to me) to have "New Tasks" in the sidebar (under the fold) and everything else that is an action button above in the title bar. (images have been submitted to the Lighthouse ticket)

I also found an inconsistency, but am unsure as to whether it is really a bug. Namely, when the user is inside a task, and is organizing, the people action says "contact organizer" instead of "edit participants". I believe this is incorrect.

I'm unsure of your process, and because I have images to show, I'm pasting this description in the lighthouse comments as well. Sorry for repeating myself!